// src/app/modules/labresults/models/labresult.ts
export interface Labresult {
  labId: string;
  accountFromID: number;
  labAccession: string;
  owner: string;
  address: string;
  contactNumber: string;
  sender: string;
  senderAddress: string;
  email: string;
  specimen: string;
  species: string;
  breed: string;
  age: number;
  gender: string;
  testMethod: string;
  fat: string;
  recommendation: string;
  diagnosis: string;
  remarks: string;
}